requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
attrs==23.2.0
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
eventlet==0.36.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
multidict==6.0.5
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
simple-websocket==1.0.0
urllib3==2.2.1
wsproto==1.2.0
yarl==1.9.4
